+++ packages/varlet-ui/src/themes/toViewport.ts
@@ -1,4 +1,5 @@
 import { StyleVars } from '../style-provider'
+import { convert } from './convert'
 
 export interface ThemesToViewportOptions {
   viewportWidth?: number
@@ -9,14 +10,5 @@ export interface ThemesToViewportOptions {
 export function toViewport(theme: StyleVars, options: ThemesToViewportOptions = {}) {
   const { viewportWidth = 375, viewportUnit = 'vmin', unitPrecision = 6 } = options
 
-  return Object.entries(theme).reduce((target, [key, value]) => {
-    target[key] = value.includes('px')
-      ? value.replace(
-          /(\d+(\.\d+)?)px/g,
-          (_, p1) => `${Number(((p1 / viewportWidth) * 100).toFixed(unitPrecision))}${viewportUnit}`,
-        )
-      : value
-
-    return target
-  }, {} as StyleVars)
+  return convert(theme, (value) => `${Number(((value / viewportWidth) * 100).toFixed(unitPrecision))}${viewportUnit}`)
 }